+++ src/exactly_lib/default/program_modes/test_case/builtin_symbols/__init__.py
@@ -1,3 +1,7 @@
-from exactly_lib.default.program_modes.test_case.builtin_symbols import test_case_dir_symbols, string_transformers
+from exactly_lib.default.program_modes.test_case.builtin_symbols import test_case_dir_symbols, strings, \
+    string_transformers
 
-ALL = test_case_dir_symbols.ALL + string_transformers.ALL
+ALL = (strings.all_strings() +
+       test_case_dir_symbols.ALL +
+       string_transformers.ALL
+       )